requirements.txt
asgiref==3.5.0
beautifulsoup4==4.10.0
coverage==6.3.1
Django==3.2.12
django-bootstrap4==21.2
django-crispy-forms==1.14.0
importlib-metadata==2.1.3
Pillow==9.0.1
pkg_resources==0.0.0
psycopg2-binary==2.8.6
python-dotenv==0.19.2
pytz==2021.3
soupsieve==2.3.1
sqlparse==0.4.2
typing_extensions==4.1.0
zipp==3.7.0
